+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v1.0.2
-- Fixed multiplayer sync (RNG seed should match between players now)
+- Fixed multiplayer sync (RNG seed should match between players now, for real)
+- Default music is now re-cached at the start of each new day (for better compatibility with modded dropship music)
 # v1.0.1
 - Increased the volume of the old music
 - You can now hear the old music while inside the building (like the other tracks)